requirements.txt
pypresence>=4.3.0
PyYAML>=6.0.2
requests>=2.32.3
python-a2s>=1.4.1
